backend/requirements.txt
Django==4.0.4
django-dotenv==1.4.2
djangorestframework==3.13.1
django-cors-headers==3.11.0
requests==2.27.1
Pillow==9.1.1
django-storages==1.12.3
boto3==1.23.5
psycopg2==2.9.3
PyYAML==6.0
Faker==13.3.4
